aes: simplify `soft` module structure
Moves the impl macros to the `soft` module
